pyproject.toml
[project]
name = "video2doc"
version = "0.1.0"
description = "A tool for transcribing videos to Markdown documents using Whisper AI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-whisper>=20250625",
    "torch>=2.0",
]

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu126" }

